채팅 카드 콤포넌트 분리
